build: these out of memory errors are killing me
This essentially overrides the default amount of RAM allocated to Java. Most modern systems have a capacity greater than 1gb believe it or not. I know Google's engineers may take a year to prepare the next version of Android on PowerMac G5s but for the rest of us, this doesn't make any sense, so let's fix it permanently.

The commits Dennis pushed earlier worked for me however Stephen's commits changed a bit so rather than rebasing, let's just do this

Change-Id: Ic9943b2d09b3ea5916f9026784e794ca3d29a1bb